Move url checker trait to UrlChecker namespace.
